Adds nested audit logs
