+++ keystone_api/apps/logging/nested.py
@@ -1,0 +1,29 @@
+from drf_spectacular.utils import extend_schema_field
+from rest_framework import serializers
+
+from apps.logging.models import *
+from apps.users.nested import UserSummarySerializer
+
+__all__ = [
+    'AuditLogSummarySerializer',
+]
+
+
+class AuditLogSummarySerializer(serializers.ModelSerializer):
+    """Object serializer for the `AuditLog` class."""
+
+    _actor = UserSummarySerializer(source='user', read_only=True)
+    action = serializers.SerializerMethodField()
+
+    class Meta:
+        """Serializer settings."""
+
+        model = AuditLog
+        fields = ['id', 'object_pk', 'object_id', 'action', 'changes', 'timestamp', 'actor', '_actor']
+
+    @extend_schema_field(str)
+    def get_action(self, obj: AuditLog) -> str:
+        """Return the logged user action as a human-readable string."""
+
+        _, as_string = AuditLog.Action.choices[obj.action]
+        return as_string
