+++ keystone_api/apps/logging/serializers.py
@@ -6,10 +6,9 @@ They encapsulate object serialization, data validation, and database object
 creation.
 """
 
-from drf_spectacular.utils import extend_schema_field
 from rest_framework import serializers
 
-from apps.users.serializers.nested import UserSummarySerializer
+from apps.users.nested import UserSummarySerializer
 from .models import *
 
 __all__ = [
@@ -18,6 +17,8 @@ __all__ = [
     'RequestLogSerializer',
     'TaskResultSerializer',
 ]
+
+from .nested import AuditLogSummarySerializer
 
 
 class AppLogSerializer(serializers.ModelSerializer):
@@ -52,21 +53,11 @@ class TaskResultSerializer(serializers.ModelSerializer):
         fields = '__all__'
 
 
-class AuditLogSerializer(serializers.ModelSerializer):
+class AuditLogSerializer(AuditLogSummarySerializer):
     """Object serializer for the `AuditLog` class."""
-
-    _actor = UserSummarySerializer(source='user', read_only=True)
-    action = serializers.SerializerMethodField()
 
     class Meta:
         """Serializer settings."""
 
         model = AuditLog
         fields = ['id', 'object_pk', 'object_id', 'action', 'changes', 'cid', 'remote_addr', 'remote_port', 'timestamp', 'actor', '_actor']
-
-    @extend_schema_field(str)
-    def get_action(self, obj: AuditLog) -> str:
-        """Return the logged user action as a human-readable string."""
-
-        _, as_string = AuditLog.Action.choices[obj.action]
-        return as_string
