+++ keystone_api/apps/users/nested.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from ..models import *
+from .models import *
 
 __all__ = [
     'TeamRoleSerializer',
