--- keystone_api/apps/users/serializers/__init__.py
@@ -1,11 +1,0 @@
-"""Serializers for casting database models to/from JSON and XML representations.
-
-Serializers handle the casting of database models to/from HTTP compatible
-representations in a manner that is suitable for use by RESTful endpoints.
-They encapsulate object serialization, data validation, and database object
-creation.
-"""
-
-from .common import *
-
-__all__ = common.__all__
